Safebox Automaton v.1.0.0
The first simulation and DFA automata generation works in console with the necessary implamantations on classes. The script allows you to generate a password for the safe box and then simulates the movements on it based on the generated automata. It needs the validation of a correct syntaz on the password